+++ petuhov.kirill/T1/circle.hpp
@@ -8,17 +8,17 @@ namespace petuhov
   class Circle : public Shape
   {
   public:
-    Circle(const point_t &center, float radius);
+    Circle(const point_t &center, double radius);
 
-    float getArea() const override;
+    double getArea() const override;
     rectangle_t getFrameRect() const override;
     void move(const point_t &newPos) override;
-    void move(float dx, float dy) override;
-    void scale(const petuhov::point_t &center, float factor) override;
+    void move(double dx, double dy) override;
+    void scale(const petuhov::point_t &center, double factor) override;
 
   private:
     point_t center_;
-    float radius_;
+    double radius_;
   };
 }
