+++ petuhov.kirill/T1/ring.cpp
@@ -2,7 +2,7 @@
 #include <stdexcept>
 #include <cmath>
 
-petuhov::Ring::Ring(const point_t &center, float outer_radius, float inner_radius):
+petuhov::Ring::Ring(const point_t &center, double outer_radius, double inner_radius):
   center_(center),
   outer_radius_(outer_radius),
   inner_radius_(inner_radius)
@@ -13,14 +13,14 @@ petuhov::Ring::Ring(const point_t &center, float outer_radius, float inner_radiu
   }
 }
 
-float petuhov::Ring::getArea() const
+double petuhov::Ring::getArea() const
 {
-  return static_cast<float>(M_PI * (outer_radius_ * outer_radius_ - inner_radius_ * inner_radius_));
+  return static_cast<double>(M_PI * (outer_radius_ * outer_radius_ - inner_radius_ * inner_radius_));
 }
 
 petuhov::rectangle_t petuhov::Ring::getFrameRect() const
 {
-  float diameter = outer_radius_ * 2;
+  double diameter = outer_radius_ * 2;
   return {{center_.x, center_.y}, diameter, diameter};
 }
 
@@ -29,21 +29,21 @@ void petuhov::Ring::move(const point_t &newPos)
   center_ = newPos;
 }
 
-void petuhov::Ring::move(float dx, float dy)
+void petuhov::Ring::move(double dx, double dy)
 {
   center_.x += dx;
   center_.y += dy;
 }
 
-void petuhov::Ring::scale(const petuhov::point_t &scale_center, float factor)
+void petuhov::Ring::scale(const petuhov::point_t &scale_center, double factor)
 {
   if (factor <= 0)
   {
     throw std::invalid_argument("Scale factor must be positive.");
   }
 
-  float dx = center_.x - scale_center.x;
-  float dy = center_.y - scale_center.y;
+  double dx = center_.x - scale_center.x;
+  double dy = center_.y - scale_center.y;
 
   outer_radius_ *= factor;
   inner_radius_ *= factor;
