+++ petuhov.kirill/T1/shape.hpp
@@ -9,11 +9,11 @@ namespace petuhov
     public:
       virtual ~Shape() = default;
 
-      virtual float getArea() const = 0;
+      virtual double getArea() const = 0;
       virtual rectangle_t getFrameRect() const = 0;
       virtual void move(const petuhov::point_t & newPos) = 0;
-      virtual void move(float dx, float dy) = 0;
-      virtual void scale(const petuhov::point_t &center, float factor) = 0;
+      virtual void move(double dx, double dy) = 0;
+      virtual void scale(const petuhov::point_t &center, double factor) = 0;
   };
 }
